+++ kwcoco/cli/__main__.py
@@ -16,6 +16,7 @@ def main(cmdline=True, **kw):
     from kwcoco.cli import coco_conform
     from kwcoco.cli import coco_modify_categories
     from kwcoco.cli import coco_reroot
+    from kwcoco.cli import coco_validate
 
     # Create a list of all submodules with CLI interfaces
     cli_modules = [
@@ -28,6 +29,7 @@ def main(cmdline=True, **kw):
         coco_conform,
         coco_eval,
         coco_modify_categories,
+        coco_validate,
     ]
 
     # Create a subparser that uses the first positional argument to run one of